README BLE Bitmap by Union
BLE Bitmap construction after receiving message from Client BLE (Smart Phone). This is one of the other functions used in my BLE communication.

The Mobile Phone sends packet to the BLE device and the BLE forwards the message to the MCU, the MCU filters the Data from Status in the packet then it organize the Data bits according to their pin functions, as shown in the code. Once organized then the MCU executes the command per the Mobile Phone request.